First program finally working, power issues were the trouble.
So simple flashing of an LED now working, turns out the code was
okay but the breadboard is one of those long ones which actually
has separate power rails in each half, and I was hooking up on
one side and thinking both sides must be powered, facepalm!

Once power was rewired, the whole thing worked fine.
NB as well as this cmd line avr dude version, Im also doing in
parallel a MicrochipStudio version which is also working! The only
difference is that in MicrochipStudio version there is a library
for delay so dont need to have a function for delay like here...